Fixed db bug - prompt buttons
When a button is clicked, it will check whether a db exists or not. And then make an API call to the relevant function with the query and isHidden set to true.

This ensures that when the messages are mapped over, the prompts from the buttons are hidden always.

Works locally. Needs testing as actual azure web app